Splits out a VisualModel base class
